requirements.txt
coverage
flake8
func-timeout
mypy
pdoc
pyo
pytest
pytest-cov
pytest-watch
python-configuration[yaml]
pyyaml
types-PyYAML
twine
wxPython